#17 #16 Add straeming start and closing jobs
This commit moves the opening of the min to min trading from being a hpsted service to being a job. This is done because we want to run the starting of the min to min trading at the beginning of a trading day. This then allows us to run a new job at the end of the trading day to close out all of our web sockets and sell any of our positions.

We are doing this because the web sockets go bad over time, so even if it was running smoothly one day, it would no longer be able to run the next. Plus, we get new stock data after hours but cant trade at that time, so there is no need to get data during that time.