Remove ensure deleted before ensure created
Doing this to try and fix the deployment on DO. I saw somewher that this can happen if too many operations appen in quick succession. Since the app crashes imdeatly on start up, I am hoping this will fix it.

There is a potential to cause an error if the following happens: the app is using SQLite and sig killed to not allow graceful shutdown. This can leave behind a DB file that will be of a cetain state. If the next time the app is booted up there was a change to the DB schema in the app, the context would fail because the old db file left over by the sig kill would be picked up. This is only an issue if the file storage under the app is non ephmeral, which, hopefully it is because it should be running on docker.

Also adds quite a bit of logging to try and narrow down where the error is if it is not this.